Move src/lib -> src/utils
