+++ src/key-map.ts
@@ -1,7 +1,7 @@
-import { get, set } from './lib/objects';
-import { firstResult } from './lib/arrays';
-import { assert } from './lib/assert';
-import { Dict } from './lib/dict';
+import { get, set } from './utils/objects';
+import { firstResult } from './utils/arrays';
+import { assert } from './utils/assert';
+import { Dict } from './utils/dict';
 import { Record } from './record';
 
 export default class KeyMap {
