+++ src/source-decorators/queryable.ts
@@ -1,5 +1,5 @@
-import { assert } from '../lib/assert';
-import { extend } from '../lib/objects';
+import { assert } from '../utils/assert';
+import { extend } from '../utils/objects';
 import { settleInSeries, fulfillInSeries } from '../evented';
 import Query from '../query';
 import Source from '../source';
