+++ src/source-decorators/updatable.ts
@@ -1,6 +1,6 @@
 import Orbit from '../main';
-import { assert } from '../lib/assert';
-import { extend } from '../lib/objects';
+import { assert } from '../utils/assert';
+import { extend } from '../utils/objects';
 import { settleInSeries, fulfillInSeries } from '../evented';
 import { Operation } from '../operation';
 import Transform from '../transform';
