+++ src/utils/dict.d.ts
